Update renpy-6.17.3-ubuntu.sh
Achei o que me fez fazer o upload do bz2, neh xcesar! Vou colocar essa extensão no .gitignore.

+++ renpy-6.17.3-ubuntu.sh
@@ -4,9 +4,30 @@ RENPY="renpy-"$VERSION"-sdk"
 BASEDIR=`pwd`
 GAME_NAME="esofteacher"
 
-# TO do: test if exists before install
-sudo apt-get install libstdc++6:i386 libgcc1:i386 zlib1g:i386 libncurses5:i386
-sudo apt-get install libsdl1.2debian:i386
+# Tests if libstdc++6:i386 exists before attempting to install.
+if ! dpkg -s libstdc++6:i386 > /dev/null; then
+    sudo apt-get install libstdc++6:i386
+fi
+
+# Tests if libgcc1:i386 exists before attempting to install.
+if ! dpkg -s libgcc1:i386 > /dev/null; then
+    sudo apt-get install libgcc1:i386
+fi
+
+# Tests if zlib1g:i386 exists before attempting to install.
+if ! dpkg -s zlib1g:i386 > /dev/null; then
+    sudo apt-get install zlib1g:i386
+fi
+
+# Tests if libncurses5:i386 exists before attempting to install.
+if ! dpkg -s libncurses5:i386 > /dev/null; then
+    sudo apt-get install libncurses5:i386
+fi
+
+# Tests if libsdl1.2debian:i386 exists before attempting to install.
+if ! dpkg -s libsdl1.2debian:i386 > /dev/null; then
+    sudo apt-get install libsdl1.2debian:i386
+fi
 
 # Tests if wxpython exists before attempting to install.
 if ! dpkg -s python-wxgtk2.8 > /dev/null; then
